pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes-web-panel"
version = "0.1.0"
description = "Standalone Python-first web panel for Hermes Agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115,<1",
  "uvicorn[standard]>=0.30,<1",
  "pydantic>=2.7,<3",
  "PyYAML>=6,<7",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["apps/server/src", "packages"]

[tool.pytest.ini_options]
pythonpath = [
  "apps/server/src",
  "packages",
]
testpaths = [
  "apps/server/tests",
  "packages",
]
